[LoRA] Keep shared Cutlass FP4 MoE path unchanged

+++ python/sglang/srt/layers/quantization/modelopt_quant.py
@@ -2222,6 +2222,6 @@ class ModelOptNvFp4FusedMoEMethod(FusedMoEMethodBase):
             params=layer.cutlass_moe_params,
             apply_router_weight_on_input=moe_runner_config.apply_router_weight_on_input,
             no_combine=moe_runner_config.no_combine,
-        )
+        ).to(x.dtype)
         # Scale by routed_scaling_factor is fused into select_experts.
         return StandardCombineInput(hidden_states=output)
